commnet function fully implemented
